backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.10.0
aiofiles>=23.2.0
networkx>=3.0
google-genai>=1.0.0
google-adk>=1.0.0
orjson>=3.9.0
json-repair>=0.7.0
loguru>=0.7.0
PyYAML>=6.0
duckduckgo-search>=4.0
httpx>=0.25.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
aiosqlite
pytest>=7.0.0
pytest-asyncio>=0.23.0
